+++ src/primitive/mod.rs
@@ -116,7 +116,6 @@ impl fmt::Display for Primitive {
                 InvTranspose => write!(f, "⍘{Transpose}"),
                 InverseBits => write!(f, "⍘{Bits}"),
                 InvTrace => write!(f, "⍘{Trace}"),
-                InvConstant => write!(f, "⍘{Constant}"),
                 Uncouple => write!(f, "⍘{Couple}"),
                 Untake => write!(f, "⍘{Take}"),
                 Undrop => write!(f, "⍘{Drop}"),
@@ -194,8 +193,8 @@ impl Primitive {
             Unroll => Roll,
             Trace => InvTrace,
             InvTrace => Trace,
-            Constant => InvConstant,
-            InvConstant => Constant,
+            Box => Unbox,
+            Unbox => Box,
             _ => return None,
         })
     }
@@ -355,12 +354,12 @@ impl Primitive {
             Primitive::Member => env.dyadic_rr_env(Value::member)?,
             Primitive::Find => env.dyadic_rr_env(Value::find)?,
             Primitive::IndexOf => env.dyadic_rr_env(Value::index_of)?,
-            Primitive::Constant => {
+            Primitive::Box => {
                 let val = env.pop(1)?;
                 let constant = Function::constant(val);
                 env.push(constant);
             }
-            Primitive::InvConstant => {
+            Primitive::Unbox => {
                 let mut val = env.pop(1)?;
                 if let Some(con) = val.as_function().and_then(|f| f.as_constant()) {
                     val = con.clone()